feat: Move to new Backend
BREAKING CHANGE: This commit adjusts the API to the current API used on the website. The parameters to specify, such as `club_id` or `category_id` may have changed. Also, as of now, only node.js v18 and above is supported, as [`fetch()`](https://nodejs.org/api/globals.html#fetch) is now used.